+++ packages/docs/site/src/routes/components/_examples/Choropleth_props.js
@@ -1,1 +1,0 @@
-export {default as worldGeojson} from '../../../../../../components/Choropleth/data/2/world_110m_iso_a2_geo.json';