+++ packages/tools/atlas/bin/cleanup.sh
@@ -1,0 +1,5 @@
+#! /usr/bin/env bash
+
+rimraf \
+world_110m_iso_a2_topo.json \
+world_110m_iso_a2_geo.json