+++ packages/tools/atlas/bin/dist.sh
@@ -1,0 +1,4 @@
+#! /usr/bin/env bash
+
+cp data/1/world_110m_iso_a2_topo.json . & \
+cp data/2/world_110m_iso_a2_geo.json .