+++ packages/tools/atlas/bin/process.sh
@@ -1,0 +1,4 @@
+#! /usr/bin/env bash
+
+bin/01_world_atlas_ids.js &&
+bin/12_world_110m_iso_a2_geojson.js